// src/types/requestCriterias.ts
export enum RessourceType {
  REQUEST = 'Request',
  IPP_LIST = 'IPPList',
  PATIENT = 'Patient',
  ENCOUNTER = 'Encounter',
  DOCUMENTS = 'DocumentReference',
  PMSI = 'pmsi',
  CONDITION = 'Condition',
  PROCEDURE = 'Procedure',
  CLAIM = 'Claim',
  MEDICATION = 'Medication',
  MEDICATION_REQUEST = 'MedicationRequest',
  MEDICATION_ADMINISTRATION = 'MedicationAdministration',
  BIO_MICRO = 'biologie_microbiologie',
  OBSERVATION = 'Observation',
  MICROBIOLOGIE = 'microbiologie',
  PHYSIOLOGIE = 'physiologie',
  IMAGING = 'ImagingStudy'
}

export enum RessourceTypeLabels {
  REQUEST = 'Mes requềtes',
  IPP_LIST = "Liste d'IPP",
  PATIENT = 'Démographie',
  ENCOUNTER = 'Prise en charge',
  DOCUMENTS = 'Documents cliniques',
  PMSI = 'PMSI',
  CONDITION = 'Diagnostic',
  PROCEDURE = 'actes CCAM',
  CLAIM = 'GHM',
  MEDICATION = 'Médicament',
  BIO_MICRO = 'Biologie/Microbiologie',
  OBSERVATION = 'Biologie',
  MICROBIOLOGIE = 'Microbiologie',
  PHYSIOLOGIE = 'Physiologie',
  IMAGING = 'Imagerie'
}

export type CommonCriteriaDataType = {
  id: number
  error?: boolean
  type: RessourceType
  encounterService?: ScopeTreeRow[]
  isInclusive?: boolean
  title: string
}
// ...
export type DocType = {
  code: string
  label: string
  type: string
}

export type DocumentDataType = CommonCriteriaDataType & {
  type: RessourceType.DOCUMENTS
  search: string
  searchBy: SearchByTypes.TEXT | SearchByTypes.DESCRIPTION
  docType: DocType[] | null
  encounterEndDate: string | null
  encounterStartDate: string | null
  occurrence: number
  occurrenceComparator: Comparators
  startOccurrence: string | null
  endOccurrence: string | null
}

export type GhmDataType = CommonCriteriaDataType & {
  type: RessourceType.CLAIM
  code: LabelObject[] | null
  encounterEndDate: string | null
  encounterStartDate: string | null
  occurrence: number
  occurrenceComparator: Comparators
  label: undefined
  startOccurrence: string | null
  endOccurrence: string | null
}
export enum Comparators {
  LESS_OR_EQUAL = '<=',
  LESS = '<',
  EQUAL = '=',
  GREATER = '>',
  GREATER_OR_EQUAL = '>=',
  BETWEEN = '<x>'
}